requirements.txt
amqp==2.3.2
beautifulsoup4==4.7.1
billiard==3.5.0.4
celery==4.2.0
certifi==2018.4.16
chardet==3.0.4
defusedxml==0.5.0
Django==2.0.1
django-celery-beat==1.1.1
django-celery-results==1.0.1
django-ckeditor==5.4.0
django-cleanup==2.1.0
django-cors-headers==2.5.2
django-filter==2.1.0
django-js-asset==1.1.0
django-widget-tweaks==1.4.2
djangorestframework==3.9.1
gunicorn==19.7.1
idna==2.7
kombu==4.2.1
lxml==4.3.2
Markdown==3.0.1
oauthlib==2.1.0
Pillow==5.1.0
psycopg2==2.7.4
psycopg2-binary==2.7.5
PyJWT==1.6.4
python3-openid==3.1.0
pytz==2018.4
redis==2.10.6
requests==2.19.1
requests-oauthlib==1.0.0
selenium==3.141.0
six==1.11.0
soupsieve==1.8
Unidecode==1.0.23
urllib3==1.23
uWSGI==2.0.17
vine==1.1.4
